perf: offload massive foliage matrix initialization to Web Workers
Extracted the heavy procedural generation logic for the GroundCover (100k blades) and BambooForest (15k stalks, branches, and leaves) out of the main thread and into dedicated Web Workers.

The previous implementation relied on running complex `SimplexNoise` loops and Three.js `Object3D` matrix transformations synchronously during React component mounting. This blocked the main thread and caused massive garbage collection (GC) stalls and slow reconciliation times.

By using Web Workers, the heavy lifting is handled entirely in parallel. The results are passed back to the main thread via zero-copy `Transferable` `ArrayBuffers`, virtually eliminating the UI freeze during load while fulfilling the "beast mode" performance mandate for high-density instanced meshes.
